[web] ♻️ refactor: move useAudioContext to global hooks folder
