+++ apps/web/src/ui/challenging/widgets/slots/ChallengeCodeEditor/useChallengeCodeEditorSlot.ts
@@ -10,9 +10,9 @@ import { useToastContext } from '@/ui/global/contexts/ToastContext'
 import { useCodeRunner } from '@/ui/global/hooks/useCodeRunner'
 import type { ConsoleRef } from '@/ui/global/widgets/components/Console/types'
 import type { CodeEditorRef } from '@/ui/global/widgets/components/CodeEditor/types'
-import { useAudioContext } from '@/ui/global/contexts/AudioContext'
 import { useLocalStorage } from '@/ui/global/hooks/useLocalStorage'
 import { useRouter } from '@/ui/global/hooks/useRouter'
+import { useAudioContext } from '@/ui/global/hooks/useAudioContext'
 
 export function useChallengeCodeEditorSlot() {
   const { getChallengeSlice, getPanelsLayoutSlice, getResultsSlice } = useChallengeStore()
