--- apps/web/src/ui/global/contexts/AudioContext/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useAudioContext } from './useAudioContext'
-export { useAudioProvider } from './useAudioProvider'
