+++ apps/web/src/ui/global/hooks/useAudioContext.ts
@@ -1,9 +1,7 @@
-'use client'
-
 import { useContext } from 'react'
 
 import { AppError } from '@stardust/core/global/errors'
-import { AudioContext } from '..'
+import { AudioContext } from '../contexts/AudioContext'
 
 export function useAudioContext() {
   const context = useContext(AudioContext)
