+++ apps/web/src/ui/global/widgets/components/PlaygroundCodeEditor/usePlaygroundCodeEditor.ts
@@ -7,7 +7,7 @@ import { Code } from '@stardust/core/global/structures'
 import { REGEX } from '@/constants'
 import { useToastContext } from '@/ui/global/contexts/ToastContext'
 import { useCodeRunner } from '@/ui/global/hooks/useCodeRunner'
-import { useAudioContext } from '@/ui/global/contexts/AudioContext'
+import { useAudioContext } from '@/ui/global/hooks/useAudioContext'
 import type { ConsoleRef } from '../Console/types'
 import type { PromptRef } from '../Prompt/types'
 import type { CodeEditorRef } from '../CodeEditor/types'
