+++ apps/web/src/ui/lesson/widgets/pages/Rewarding/useRewardingPage.ts
@@ -5,7 +5,7 @@ import { type RefObject, useCallback, useEffect, useState } from 'react'
 import type { AlertDialogRef } from '@/ui/global/widgets/components/AlertDialog/types'
 import { useRouter } from '@/ui/global/hooks/useRouter'
 import { useRefreshPage } from '@/ui/global/hooks/useRefreshPage'
-import { useAudioContext } from '@/ui/global/contexts/AudioContext'
+import { useAudioContext } from '@/ui/global/hooks/useAudioContext'
 import { useCookieActions } from '@/ui/global/hooks/useCookieActions'
 import { useAuthContext } from '@/ui/auth/contexts/AuthContext'
 import { COOKIES } from '@/constants'
